Consolidates all services in one since all calling same database service
