+++ CanWeFixIt/CanWeFixIt.Api/Services/Interfaces/ICanWeFixItService.cs
@@ -4,8 +4,10 @@ using System.Threading.Tasks;
 
 namespace CanWeFixIt.Api.Services.Interfaces
 {
-    public interface IValuationService
+    public interface ICanWeFixItService
     {
+        Task<IEnumerable<Instrument>> GetInstrumentsAsync(bool? active = null);
+        Task<IEnumerable<MarketDataDto>> GetMarketDataAsync(bool? active = null);
         Task<IEnumerable<MarketValuation>> GetMarketValuationsAsync(bool? active = null);
     }
 }
