--- CanWeFixIt/CanWeFixIt.Api/Services/Interfaces/IInstrumentService.cs
@@ -1,11 +1,0 @@
-﻿using CanWeFixIt.Data.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace CanWeFixIt.Api.Services.Interfaces
-{
-    public interface IInstrumentService
-    {
-        Task<IEnumerable<Instrument>> GetInstrumentsAsync(bool? active = null);
-    }
-}
