--- CanWeFixIt/CanWeFixIt.Api/Services/Interfaces/IMarketDataService.cs
@@ -1,11 +1,0 @@
-﻿using CanWeFixIt.Data.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace CanWeFixIt.Api.Services.Interfaces
-{
-    public interface IMarketDataService
-    {
-        Task<IEnumerable<MarketDataDto>> GetMarketDataAsync(bool? active = null);
-    }
-}
